requirements.txt
streamlit
pymupdf
langdetect
python-dotenv	
sentence-transformers
langchain-community
transformers
faiss-cpu
tqdm
langchain-groq
langchain